+++ CHANGELOG.md
@@ -1,9 +1,12 @@
-# RUF r203-Alpha
+# RUF r204-Alpha
 ### New
-* Added Show in raid option for Party & Party Pets to enable showing these units in a raid group,
+* Combo Points supported in Classic under Class Power options.
+* Added Show in raid option for Party & Party Pets to enable showing these units in a raid group.
 * Added Party Pets as a supported Unit. Default options need work.
 
 ### Fixed
+* Castbars should properly enable and disable themselves now.
+* Buffs and Debuffs should properly when one is disabled. (for real this time)
 * Range Fading should be far more accurate for assistable, friendly units (i.e party members or friendly NPCs that you can buff or heal).
 
 ### Known Issues